+++ src/message_log.h
@@ -14,8 +14,8 @@ public:
     std::vector<std::string> system;      // active on-screen messages
     std::vector<std::string> history;     // persistent log
     bool is_active = false;              // messages pending display
-    unsigned int max_len_msg = 0;        // longest message char count
-    unsigned int max_len_msg_measure = 0; // pixel width of longest
+    size_t max_len_msg = 0;        // longest message char count
+    int max_len_msg_measure = 0;   // pixel width of longest
 
     bool add(const char* fmt, ...);
     void add_history(const char* fmt, ...);
